+++ backend/app/main.py
@@ -1,62 +1,13 @@
-from fastapi import Depends, FastAPI, HTTPException, Query
+from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.routing import APIRoute
-from prisma.models import (
-    AccessLevel as DbAccessLevel,
-)
-from prisma.models import (
-    Course as DbCourse,
-)
-from prisma.models import (
-    Major as DbMajor,
-)
-from prisma.models import (
-    Minor as DbMinor,
-)
-from prisma.models import (
-    Title as DbTitle,
-)
-from prisma.types import CourseWhereInput, CourseWhereInputRecursive2
-from pydantic import BaseModel
-from unidecode import unidecode
 
-from . import sync
+from . import routes, sync
 from .database import prisma
 from .plan.courseinfo import (
-    CourseDetails,
-    EquivDetails,
-    clear_course_info_cache,
     course_info,
-    make_searchable_name,
 )
-from .plan.generation import generate_empty_plan, generate_recommended_plan
-from .plan.plan import ValidatablePlan
-from .plan.storage import (
-    LowDetailPlanView,
-    PlanView,
-    get_plan_details,
-    get_user_plans,
-    modify_plan_metadata,
-    modify_validatable_plan,
-    remove_plan,
-    store_plan,
-)
-from .plan.validation.curriculum.solve import solve_curriculum
-from .plan.validation.diagnostic import ValidationResult
-from .plan.validation.validate import diagnose_plan
-from .sync.siding import translate as siding_translate
 from .sync.siding.client import client as siding_soap_client
-from .user.auth import (
-    AccessLevelOverview,
-    AdminKey,
-    ModKey,
-    UserKey,
-    login_cas,
-    require_admin_auth,
-    require_authentication,
-    require_mod_auth,
-)
-from .user.info import StudentContext
 
 
 # Set-up operation IDs for OpenAPI
@@ -110,578 +61,5 @@ async def health():
     return {"message": "OK"}
 
 
-@app.get("/auth/login")
-async def authenticate(
-    next: str | None = None,
-    ticket: str | None = None,
-):
-    """
-    Redirect the browser to this page to initiate authentication.
-    """
-    return await login_cas(next, ticket)
-
-
-@app.get("/auth/check")
-async def check_auth(user: UserKey = Depends(require_authentication)):
-    """
-    Request succeeds if user authentication was successful.
-    Otherwise, the request fails with 401 Unauthorized.
-    """
-    return {"message": "Authenticated"}
-
-
-@app.get("/auth/check/mod")
-async def check_mod(user: ModKey = Depends(require_mod_auth)):
-    """
-    Request succeeds if user authentication and mod authorization were successful.
-    Otherwise, the request fails with 401 Unauthorized or 403 Forbidden.
-    """
-    return {"message": "Authenticated with mod access"}
-
-
-@app.get("/auth/check/admin")
-async def check_admin(user: AdminKey = Depends(require_admin_auth)):
-    """
-    Request succeeds if user authentication and admin authorization were successful.
-    Otherwise, the request fails with 401 Unauthorized or 403 Forbidden.
-    """
-    return {"message": "Authenticated with admin access"}
-
-
-@app.get("/auth/mod", response_model=list[AccessLevelOverview])
-async def view_mods(user: AdminKey = Depends(require_admin_auth)):
-    """
-    Show a list of all current mods with username and RUT. Up to 50 records.
-    """
-    mods = await DbAccessLevel.prisma().find_many(take=50)
-
-    named_mods: list[AccessLevelOverview] = []
-    for mod in mods:
-        named_mods.append(AccessLevelOverview(**dict(mod)))
-        try:
-            print(f"fetching user data for user {mod.user_rut} from SIDING...")
-            # TODO: check if this function works for non-students
-            data = await siding_translate.fetch_student_info(mod.user_rut)
-            named_mods[-1].name = data.full_name
-        finally:
-            # Ignore if couldn't get the name by any reason to at least show
-            # the RUT, which is more important.
-            pass
-    return named_mods
-
-
-@app.post("/auth/mod")
-async def add_mod(rut: str, user: AdminKey = Depends(require_admin_auth)):
-    """
-    Give mod access to a user with the specified RUT.
-    """
-    return await DbAccessLevel.prisma().upsert(
-        where={
-            "user_rut": rut,
-        },
-        data={
-            "create": {
-                "user_rut": rut,
-                "is_mod": True,
-            },
-            "update": {
-                "is_mod": True,
-            },
-        },
-    )
-
-
-@app.delete("/auth/mod")
-async def remove_mod(rut: str, user: AdminKey = Depends(require_admin_auth)):
-    """
-    Remove mod access from a user with the specified RUT.
-    """
-    mod_record = await DbAccessLevel.prisma().find_unique(where={"user_rut": rut})
-
-    if not mod_record:
-        raise HTTPException(status_code=404, detail="Mod not found")
-    return await DbAccessLevel.prisma().delete(where={"user_rut": rut})
-
-
-@app.get("/student/info", response_model=StudentContext)
-async def get_student_info(user: UserKey = Depends(require_authentication)):
-    """
-    Get the student info for the currently logged in user.
-    Requires authentication (!)
-    This forwards a request to the SIDING service.
-    """
-    return await sync.get_student_data(user)
-
-
-@app.post("/sync")
-async def sync_database(
-    courses: bool = False,
-    offer: bool = False,
-    admin: AdminKey = Depends(require_admin_auth),
-):
-    """
-    Initiate a synchronization of the internal database from external sources.
-    """
-    await sync.run_upstream_sync(courses, offer)
-    return {
-        "message": "Database updated from external sources",
-    }
-
-
-class CourseOverview(BaseModel):
-    code: str
-    name: str
-    credits: int
-    school: str
-    area: str | None
-    is_available: bool
-
-
-class CourseFilter(BaseModel):
-    # Only allow courses that match the given search string, in name or course code.
-    text: str | None = None
-    # Only allow courses that have the given amount of credits.
-    credits: int | None = None
-    # Only allow courses matching the given school.
-    school: str | None = None
-    # Only allow courses that match the given availability.
-    available: bool | None = None
-    # Only allow courses that are available/unavailable on first semesters.
-    first_semester: bool | None = None
-    # Only allow courses that are available/unavailable on second semesters.
-    second_semester: bool | None = None
-    # Only allow courses that are members of the given equivalence.
-    equiv: str | None = None
-
-    def as_db_filter(self) -> CourseWhereInput:
-        filter = CourseWhereInput()
-        if self.text is not None:
-            search_text = make_searchable_name(self.text)
-            name_parts: list[CourseWhereInputRecursive2] = [
-                {"searchable_name": {"contains": text_part}}
-                for text_part in search_text.split()
-            ]
-            filter["OR"] = [
-                {"code": {"contains": search_text.upper()}},
-                {"AND": name_parts},
-            ]
-        if self.credits is not None:
-            filter["credits"] = self.credits
-        if self.school is not None:
-            ascii_school = unidecode(self.school)
-            filter["school"] = {"contains": ascii_school, "mode": "insensitive"}
-        if self.available is not None:
-            filter["is_available"] = self.available
-        if self.first_semester is not None:
-            filter["semestrality_first"] = self.first_semester
-        if self.second_semester is not None:
-            filter["semestrality_second"] = self.second_semester
-        if self.equiv is not None:
-            filter["equivs"] = {"some": {"equiv_code": self.equiv}}
-        return filter
-
-
-# This should be a GET request, but FastAPI does not support JSON in GET requests
-# easily.
-# See https://github.com/tiangolo/fastapi/discussions/7919
-@app.post("/courses/search/details", response_model=list[CourseOverview])
-async def search_course_details(filter: CourseFilter):
-    """
-    Fetches a list of courses that match the given name (or code),
-    credits and school.
-    """
-    return await DbCourse.prisma().find_many(where=filter.as_db_filter(), take=50)
-
-
-# This should be a GET request, but FastAPI does not support JSON in GET requests
-# easily.
-# See https://github.com/tiangolo/fastapi/discussions/7919
-@app.post("/courses/search/codes", response_model=list[str])
-async def search_course_codes(filter: CourseFilter):
-    """
-    Fetches a list of courses that match the given name (or code),
-    credits and school.
-    Returns only the course codes, but allows up to 3000 results.
-    """
-    return [
-        c.code
-        for c in await DbCourse.prisma().find_many(
-            where=filter.as_db_filter(),
-            take=3000,
-        )
-    ]
-
-
-@app.get("/courses", response_model=list[CourseDetails])
-async def get_course_details(codes: list[str] = Query()) -> list[CourseDetails]:
-    """
-    For a list of course codes, fetch a corresponding list of course details.
-
-    Request example: `/api/courses?codes=IIC2233&codes=IIC2173`
-    """
-
-    courseinfo = await course_info()
-    courses: list[CourseDetails] = []
-    for code in codes:
-        course = courseinfo.try_course(code)
-        if course is None:
-            raise HTTPException(status_code=404, detail=f"Course '{code}' not found")
-        courses.append(course)
-    return courses
-
-
-@app.get("/equivalences", response_model=list[EquivDetails])
-async def get_equivalence_details(
-    codes: list[str] = Query(),
-) -> list[EquivDetails]:
-    """
-    For a list of equivalence codes, fetch the raw equivalence details, without any
-    filtering.
-    To filter courses for a specific equivalence, use `search_courses` with an
-    equivalence filter.
-    """
-
-    courseinfo = await course_info()
-    equivs: list[EquivDetails] = []
-    for code in codes:
-        equiv = courseinfo.try_equiv(code)
-        if equiv is None:
-            raise HTTPException(
-                status_code=404,
-                detail=f"Equivalence '{code}' not found",
-            )
-        equivs.append(equiv)
-    return equivs
-
-
-@app.post("/plan/rebuild")
-async def rebuild_validation_rules():
-    """
-    Recache course information from internal database.
-    """
-    await clear_course_info_cache()
-    info = await course_info()
-    return {
-        "message": f"Recached {len(info.courses)} courses and "
-        f"{len(info.equivs)} equivalences",
-    }
-
-
-@app.get("/plan/empty_for", response_model=ValidatablePlan)
-async def empty_plan_for_user(user: UserKey = Depends(require_authentication)):
-    """
-    Generate an empty plan using the current user as context.
-    For example, the created plan includes all passed courses, uses the curriculum
-    version for the given user and selects the student's official choice of
-    major/minor/title if available.
-
-    (Currently this is equivalent to `empty_guest_plan()` until we get user data)
-    """
-    return await generate_empty_plan(user)
-
-
-@app.get("/plan/empty_for_any", response_model=ValidatablePlan)
-async def empty_plan_for_any_user(
-    user_rut: str,
-    mod: ModKey = Depends(require_mod_auth),
-):
-    """
-    Same functionality as `empty_plan_for_user`, but works for any user identified by
-    their RUT with `user_rut`.
-    Moderator access is required.
-    """
-    return await generate_empty_plan(mod.as_any_user(user_rut))
-
-
-@app.get("/plan/empty_guest", response_model=ValidatablePlan)
-async def empty_guest_plan():
-    """
-    Generates a generic empty plan with no user context, using the latest curriculum
-    version.
-    """
-    return await generate_empty_plan(None)
-
-
-@app.post("/plan/validate", response_model=ValidationResult)
-async def validate_guest_plan(plan: ValidatablePlan):
-    """
-    Validate a plan, generating diagnostics.
-    """
-    return await diagnose_plan(plan, user_ctx=None)
-
-
-@app.post("/plan/validate_for", response_model=ValidationResult)
-async def validate_plan_for_user(
-    plan: ValidatablePlan,
-    user: UserKey = Depends(require_authentication),
-):
-    """
-    Validate a plan, generating diagnostics.
-    Includes diagnostics tailored for the given user and skips diagnostics that do not
-    apply to the particular student.
-    """
-    user_ctx = await sync.get_student_data(user)
-    return await diagnose_plan(plan, user_ctx)
-
-
-@app.post("/plan/validate_for_any", response_model=ValidationResult)
-async def validate_plan_for_any_user(
-    plan: ValidatablePlan,
-    user_rut: str,
-    mod: ModKey = Depends(require_mod_auth),
-):
-    """
-    Same functionality as `validate_plan_for_user`, but works for any user identified by
-    their RUT with `user_rut`.
-    Moderator access is required.
-    """
-    user_ctx = await sync.get_student_data(mod.as_any_user(user_rut))
-    return await diagnose_plan(plan, user_ctx)
-
-
-@app.post("/plan/curriculum_graph")
-async def get_curriculum_validation_graph(plan: ValidatablePlan) -> str:
-    """
-    Get the curriculum validation graph for a certain plan, in Graphviz DOT format.
-    Useful for debugging and kind of a bonus easter egg.
-    """
-    courseinfo = await course_info()
-    curriculum = await sync.get_curriculum(plan.curriculum)
-    g = solve_curriculum(courseinfo, curriculum, plan.classes)
-    return g.dump_graphviz()
-
-
-@app.post("/plan/generate", response_model=ValidatablePlan)
-async def generate_plan(passed: ValidatablePlan):
-    """
-    From a base plan, generate a new plan that should lead the user to earn their title
-    of choice.
-    """
-    return await generate_recommended_plan(passed)
-
-
-@app.post("/plan/storage", response_model=PlanView)
-async def save_plan(
-    name: str,
-    plan: ValidatablePlan,
-    user: UserKey = Depends(require_authentication),
-) -> PlanView:
-    """
-    Save a plan with the given name in the storage of the current user.
-    Fails if the user is not logged in.
-    """
-    return await store_plan(plan_name=name, user=user, plan=plan)
-
-
-@app.post("/plan/storage/any", response_model=PlanView)
-async def save_any_plan(
-    name: str,
-    plan: ValidatablePlan,
-    user_rut: str,
-    mod: ModKey = Depends(require_mod_auth),
-) -> PlanView:
-    """
-    Same functionality as `save_plan`, but works for any user identified by
-    their RUT with `user_rut`.
-    Moderator access is required.
-    All `/plan/storage/any` endpoints (and sub-resources) should require
-    moderator access.
-    """
-    return await store_plan(plan_name=name, user=mod.as_any_user(user_rut), plan=plan)
-
-
-@app.get("/plan/storage", response_model=list[LowDetailPlanView])
-async def read_plans(
-    user: UserKey = Depends(require_authentication),
-) -> list[LowDetailPlanView]:
-    """
-    Fetches an overview of all the plans in the storage of the current user.
-    Fails if the user is not logged in.
-    Does not return the courses in each plan, only the plan metadata required
-    to show the users their list of plans (e.g. the plan id).
-    """
-    return await get_user_plans(user)
-
-
-@app.get("/plan/storage/any", response_model=list[LowDetailPlanView])
-async def read_any_plans(
-    user_rut: str,
-    mod: ModKey = Depends(require_mod_auth),
-) -> list[LowDetailPlanView]:
-    """
-    Same functionality as `read_plans`, but works for any user identified by
-    their RUT with `user_rut`.
-    Moderator access is required.
-    """
-    return await get_user_plans(mod.as_any_user(user_rut))
-
-
-@app.get("/plan/storage/details", response_model=PlanView)
-async def read_plan(
-    plan_id: str,
-    user: UserKey = Depends(require_authentication),
-) -> PlanView:
-    """
-    Fetch the plan details for a given plan id.
-    Requires the current user to be the plan owner.
-    """
-    return await get_plan_details(user=user, plan_id=plan_id)
-
-
-@app.get("/plan/storage/any/details", response_model=PlanView)
-async def read_any_plan(
-    plan_id: str,
-    user: ModKey = Depends(require_mod_auth),
-) -> PlanView:
-    """
-    Same functionality as `read_plan`, but works for any plan of any user
-    identified by their RUT.
-    Moderator access is required.
-    """
-    return await get_plan_details(user=user, plan_id=plan_id, mod_access=True)
-
-
-@app.put("/plan/storage", response_model=PlanView)
-async def update_plan(
-    plan_id: str,
-    new_plan: ValidatablePlan,
-    user: UserKey = Depends(require_authentication),
-) -> PlanView:
-    """
-    Modifies the courses of a plan by id.
-    Requires the current user to be the owner of this plan.
-    Returns the updated plan.
-    """
-    return await modify_validatable_plan(user=user, plan_id=plan_id, new_plan=new_plan)
-
-
-@app.put("/plan/storage/any", response_model=PlanView)
-async def update_any_plan(
-    plan_id: str,
-    new_plan: ValidatablePlan,
-    user: ModKey = Depends(require_mod_auth),
-) -> PlanView:
-    """
-    Same functionality as `update_plan`, but works for any plan of any user
-    identified by their RUT.
-    Moderator access is required.
-    """
-    return await modify_validatable_plan(
-        user=user,
-        plan_id=plan_id,
-        new_plan=new_plan,
-        mod_access=True,
-    )
-
-
-@app.put("/plan/storage/metadata", response_model=PlanView)
-async def update_plan_metadata(
-    plan_id: str,
-    set_name: str | None = None,
-    set_favorite: bool | None = None,
-    user: UserKey = Depends(require_authentication),
-) -> PlanView:
-    """
-    Modifies the metadata of a plan (currently only `name` or `is_favorite`).
-    Modify one attribute per request.
-    Requires the current user to be the owner of this plan.
-    Returns the updated plan.
-    """
-
-    return await modify_plan_metadata(
-        user=user,
-        plan_id=plan_id,
-        set_name=set_name,
-        set_favorite=set_favorite,
-    )
-
-
-@app.put("/plan/storage/any/metadata", response_model=PlanView)
-async def update_any_plan_metadata(
-    plan_id: str,
-    set_name: str | None = None,
-    set_favorite: bool | None = None,
-    mod: ModKey = Depends(require_mod_auth),
-) -> PlanView:
-    """
-    Same functionality as `update_plan_metadata`, but works for any plan of any user
-    identified by their RUT.
-    Moderator access is required.
-    """
-
-    return await modify_plan_metadata(
-        user=mod,
-        plan_id=plan_id,
-        set_name=set_name,
-        set_favorite=set_favorite,
-        mod_access=True,
-    )
-
-
-@app.delete("/plan/storage", response_model=PlanView)
-async def delete_plan(
-    plan_id: str,
-    user: UserKey = Depends(require_authentication),
-) -> PlanView:
-    """
-    Deletes a plan by ID.
-    Requires the current user to be the owner of this plan.
-    Returns the removed plan.
-    """
-    return await remove_plan(user=user, plan_id=plan_id)
-
-
-@app.delete("/plan/storage/any", response_model=PlanView)
-async def delete_any_plan(
-    plan_id: str,
-    mod: ModKey = Depends(require_mod_auth),
-) -> PlanView:
-    """
-    Same functionality as `delete_plan`, but works for any plan of any user
-    identified by their RUT.
-    Moderator access is required.
-    """
-    return await remove_plan(user=mod, plan_id=plan_id, mod_access=True)
-
-
-@app.get("/offer/major", response_model=list[DbMajor])
-async def get_majors(cyear: str):
-    """
-    Get all the available majors for a given curriculum version (cyear).
-    """
-    return await DbMajor.prisma().find_many(
-        where={
-            "cyear": cyear,
-        },
-    )
-
-
-@app.get("/offer/minor", response_model=list[DbMinor])
-async def get_minors(cyear: str, major_code: str | None = None):
-    if major_code is None:
-        return await DbMinor.prisma().find_many(
-            where={
-                "cyear": cyear,
-            },
-        )
-    return await DbMinor.prisma().query_raw(
-        """
-        SELECT *
-        FROM "Minor", "MajorMinor"
-        WHERE "MajorMinor".minor = "Minor".code
-            AND "MajorMinor".major = $2
-            AND "MajorMinor".cyear = $1
-            AND "Minor".cyear = $1
-        """,
-        cyear,
-        major_code,
-    )
-
-
-@app.get("/offer/title", response_model=list[DbTitle])
-async def get_titles(cyear: str):
-    return await DbTitle.prisma().find_many(
-        where={
-            "cyear": cyear,
-        },
-    )
+for router in routes.routers:
+    app.include_router(router)
