+++ backend/app/routes/__init__.py
@@ -1,0 +1,7 @@
+"""
+Routes are grouped under this module.
+"""
+
+from . import admin, courses, offer, plan, user
+
+routers = [admin.router, courses.router, offer.router, plan.router, user.router]
